Move Checkpoint to IterationData
