+++ dali/pipeline/operator/checkpointing/checkpoint_test.cc
@@ -20,6 +20,8 @@
 
 #include "dali/test/dali_test.h"
 #include "dali/core/cuda_stream_pool.h"
+#include "dali/pipeline/operator/operator.h"
+#include "dali/pipeline/graph/op_graph.h"
 
 namespace dali {
